+++ packages/lix-sdk/src/state/commit.bench.ts
@@ -1,7 +1,7 @@
 import { bench } from "vitest";
 import { openLix } from "../lix/open-lix.js";
 import { commit } from "./commit.js";
-import { insertTransactionState } from "./insert-transaction-state.js";
+import { insertTransactionState } from "./transaction/insert-transaction-state.js";
 import { timestamp } from "../deterministic/timestamp.js";
 
 // NOTE: openLix includes database initialization overhead
