+++ packages/lix-sdk/src/state/commit.test.ts
@@ -2,7 +2,7 @@ import { test, expect, describe } from "vitest";
 import type { Kysely } from "kysely";
 import type { LixInternalDatabaseSchema } from "../database/schema.js";
 import type { LixCommitEdge } from "../commit/schema.js";
-import { insertTransactionState } from "./insert-transaction-state.js";
+import { insertTransactionState } from "./transaction/insert-transaction-state.js";
 import { commit } from "./commit.js";
 import { openLix } from "../lix/open-lix.js";
 import { nanoId, timestamp, uuidV7 } from "../deterministic/index.js";
